ifcfg-rh: use binary search for converting string to ethtool ID
Don't do a linear search through all names, but use binary search.

Upside: calling nms_ifcfg_rh_utils_get_ethtool_by_name() in a loop
(once over all 60 names) is 75% faster.

Downside: when adding a new feature, we have yet another line that we
need to add. Previously, adding a new feature required adding 7 lines,
not it is 8. But we didn't add a single feature since this was added,
so that happens very seldom.

Possible downside: is this code harder to read? That we now track both how to
convert the ID to name and also back, is redundant (and thus harder to
maintain). But it's really just one extra line per item, for which there
is a unit test. So, I don't think when adding a new NMEthtoolID that you
even have any chance to getting this wrong (because of the test). So,
maybe it's slightly harder to read. On the other hand, it unifies
handling for ethtools and kernel names, and the code has less logic
and became more descriptive. I don't think this is harder to maintain
and it should be easy to see that it is correct (readability).