ovs: set the MTU in the ovsdb
From the Open vSwitch FAQ:

Q: How can I configure the bridge internal interface MTU? Why does
   Open vSwitch keep changing internal ports MTU?

    A: By default Open vSwitch overrides the internal interfaces
    (e.g. br0) MTU. If you have just an internal interface (e.g. br0)
    and a physical interface (e.g. eth0), then every change in MTU to
    eth0 will be reflected to br0. Any manual MTU configuration using
    ip on internal interfaces is going to be overridden by Open
    vSwitch to match the current bridge minimum.

Therefore, the safest way to set the MTU for an ovs interface is to
update it directly in the ovsdb and wait that the value gets applied
by ovs-vswitchd. Once the link MTU changes, the IPv6 MTU can be
applied as well.

Note that for ovs interfaces we don't guarantee that the specified MTU
is set when activation succeeds. The reason is that ovs doesn't
provide a convenient API to do so.